Add files via upload
The UNited States of America consistenly topped the charts for all medals won (gold, silver, bronze) with a wide margin despite reporting an average population and ranking 6th position on the GDP table while Ireland on the other hand, topped the GDP table and wasnt amongst the top 10 countries on the medal tables which further explains that having a high GDP does not equate winning medals at sports or competition. The following observations and recommendations were drawn from this analysis.

OBSERVATION.

The United States of America not only dominated the sports competition but also topped the chart for the gold, silver and bronze medals 
Ireland topped the chart for the GDP revenues but was not amongst the Top !) countries for any of the medals won at the Olympics
An observation that countries who perfomed brilliantly at the competition had very good investments in their sports faciliites and infrastructures etc
however, from these observations, few recommendations were also made and they include

RECOMMENDATION.

Countries that seek to develop their sports ability can understand that it does not solely depend on the GDP but also on their interest and capabilities. with the little the offices in charge can do, develop their sports facilities, encourage younger talents through empower programmes and awareness.
Countries like Ireland who have a high GDP returns but low chances of winning can consider channeling a good percentage of this revenue into their sports department through the right authorities to encourage sports in the country, groom young minds and talents.
Upcoming countries are recommended to leverage on collaborations and partnerships with already developed sports countries with the desire to gain insight on sports training, mentorships, coaching and funding so as to get better at it. 
Orgaizing in house competitions with team members of same country is also a good way to spike interest and improve craftsmanship at sports of interest. it is also a good way to boost their overall performance and confidence at competitions.
FInally, Sports asides being a good way to generate revenue, team spirit, and recognmition for countries or parties involbved has also proven to be a good way to improve health physically and mentally for individuals that practise it.